**What problem is this PR intended to solve?**

While working with @stevecheckoway on #3205, we caught an edge case in
how the `max_depth` parameter was converted from signed to unsigned int.
It turned out that it would have been caught much earlier with the
`-Wconversion` flag, but that flag is useless in Ruby C extensions
because of the issue described at
https://bugs.ruby-lang.org/issues/20507

That issue was fixed in Ruby 3.4-dev by
https://github.com/ruby/ruby/pull/10843 and so I'm using it here and
fixing the compilation warnings it flags.


**Have you included adequate test coverage?**

N/A


**Does this change affect the behavior of either the C or the Java
implementations?**

N/A

+++ ext/nokogiri/xml_document.c
@@ -109,7 +109,7 @@ memsize_node(const xmlNodePtr node)
   xmlAttrPtr property;
   size_t memsize = 0;
 
-  memsize += xmlStrlen(node->name);
+  memsize += (size_t)xmlStrlen(node->name);
 
   if (node->type == XML_ELEMENT_NODE) {
     for (property = node->properties; property; property = property->next) {
@@ -117,7 +117,7 @@ memsize_node(const xmlNodePtr node)
     }
   }
   if (node->type == XML_TEXT_NODE) {
-    memsize += xmlStrlen(node->content);
+    memsize += (size_t)xmlStrlen(node->content);
   }
   for (child = node->children; child; child = child->next) {
     memsize += sizeof(xmlNode) + memsize_node(child);
